refactor datasets/utils to remove cross-dependencies
